+++ CHANGELOG.md
@@ -1,13 +1,21 @@
-# poloLender Pro ChangeLog
+# poloLender Pro Changelog
 
 All notable changes to this module are documented in this file.
 poloLender application adheres to [Semantic Versioning](http://semver.org/).
+
+## [0.6.1] - 2017-03-28
+
+### Added
+- API calls throttling to comply with Poloniex limits
+
+### Changed
+- update dependency on Poloniex API wrapper to correctly show API error message details
 
 ## [0.6.0] - 2017-01-24
 
 ### Added
 - support for lending Dash (DASH)
-- 
+
 ## [0.5.0] - 2016-10-11
 
 ### Added